fix hash test

+++ t/sperl_t_hash.c
@@ -35,51 +35,51 @@ int main()
   // Hash - first same key
   {
     SPerl_HASH* hash = SPerl_HASH_new(sperl, 101);
-    int32_t value1 = 3;
+    int64_t value1 = 3;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value1);
 
-    int32_t value1_ret = *(int32_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
+    int64_t value1_ret = *(int64_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
     OK(value1_ret == 3);
 
-    int32_t value2 = 5;
+    int64_t value2 = 5;
     SPerl_HASH_insert_norehash(sperl, hash, "key", 3, &value2);
     
-    int32_t value2_ret = *(int32_t*)SPerl_HASH_search(sperl, hash, "key", 3);
+    int64_t value2_ret = *(int64_t*)SPerl_HASH_search(sperl, hash, "key", 3);
     OK(value2_ret == 5);
   }
   
   // Hash - insert and search
   {
     SPerl_HASH* hash = SPerl_HASH_new(sperl, 101);
-    int32_t value1 = 3;
+    int64_t value1 = 3;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value1);
-    int32_t hash_value1 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
-    int32_t index1 = hash_value1 % 101;
-    
-    OK(*(int32_t*)hash->table[index1]->value == 3);
+    int64_t hash_value1 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
+    int64_t index1 = hash_value1 % 101;
+   
+    OK(*(int64_t*)hash->table[index1]->value == 3);
     OK(hash->entries_length == 1);
 
-    int32_t value2 = 5;
+    int64_t value2 = 5;
     SPerl_HASH_insert_norehash(sperl, hash, "key2", 4, &value2);
-    int32_t hash_value2 = SPerl_HASH_FUNC_calc_hash(sperl, "key2", 4);
-    int32_t index2 = hash_value2 % 101;
+    int64_t hash_value2 = SPerl_HASH_FUNC_calc_hash(sperl, "key2", 4);
+    int64_t index2 = hash_value2 % 101;
     
-    OK(*(int32_t*)hash->table[index2]->value == 5);
+    OK(*(int64_t*)hash->table[index2]->value == 5);
     OK(hash->entries_length == 2);
     
     // Replace
-    int32_t value3 = 7;
+    int64_t value3 = 7;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value3);
-    int32_t hash_value3 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
-    int32_t index3 = hash_value3 % 101;
+    int64_t hash_value3 = SPerl_HASH_FUNC_calc_hash(sperl, "key1", 4);
+    int64_t index3 = hash_value3 % 101;
     
-    OK(*(int32_t*)hash->table[index3]->value == 7);
+    OK(*(int64_t*)hash->table[index3]->value == 7);
     OK(hash->entries_length == 2);
     
     // Search
-    int32_t value1_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
+    int64_t value1_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
     OK(value1_1 == 7);
-    int32_t value2_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
+    int64_t value2_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
     OK(value2_1 == 5);
     void* value3_1 = SPerl_HASH_search(sperl, hash, "none", 4);
     OK(!value3_1);
@@ -89,44 +89,44 @@ int main()
   // Hash - insert link
   {
     SPerl_HASH* hash = SPerl_HASH_new(sperl, 1);
-    int32_t value1 = 3;
+    int64_t value1 = 3;
     SPerl_HASH_insert_norehash(sperl, hash, "key1", 4, &value1);
     
-    OK(*(int32_t*)hash->table[0]->value == 3);
+    OK(*(int64_t*)hash->table[0]->value == 3);
     OK(hash->entries_length == 1);
 
-    int32_t value2 = 5;
+    int64_t value2 = 5;
     SPerl_HASH_insert_norehash(sperl, hash, "key2", 4, &value2);
     
-    OK(*(int32_t*)hash->table[0]->next->value == 5);
+    OK(*(int64_t*)hash->table[0]->next->value == 5);
     OK(hash->entries_length == 1);
     
-    int32_t value3 = 7;
+    int64_t value3 = 7;
     SPerl_HASH_insert_norehash(sperl, hash, "key3", 4, &value3);
     
-    OK(*(int32_t*)hash->table[0]->next->next->value == 7);
+    OK(*(int64_t*)hash->table[0]->next->next->value == 7);
     OK(hash->entries_length == 1);
     
     // Replace key2
-    int32_t value4 = 11;
+    int64_t value4 = 11;
     SPerl_HASH_insert_norehash(sperl, hash, "key2", 4, &value4);
     
-    OK(*(int32_t*)hash->table[0]->next->value == 11);
+    OK(*(int64_t*)hash->table[0]->next->value == 11);
     OK(hash->entries_length == 1);
     
     // Replace key3
-    int32_t value5 = 13;
+    int64_t value5 = 13;
     SPerl_HASH_insert_norehash(sperl, hash, "key3", 4, &value5);
     
-    OK(*(int32_t*)hash->table[0]->next->next->value == 13);
+    OK(*(int64_t*)hash->table[0]->next->next->value == 13);
     OK(hash->entries_length == 1);
 
     // Search
-    int32_t value1_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
+    int64_t value1_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key1", 4);
     OK(value1_1 == 3);
-    int32_t value2_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
+    int64_t value2_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key2", 4);
     OK(value2_1 == 11);
-    int32_t value3_1 = *(int32_t*)SPerl_HASH_search(sperl, hash, "key3", 4);
+    int64_t value3_1 = *(int64_t*)SPerl_HASH_search(sperl, hash, "key3", 4);
     OK(value3_1 == 13);
     
     // free
